키보드오류 해결
